ファapp配下のファのディレクトリ名を大文字始まりにしたのとサービスじゃServiceじゃなくてModelにした
