+++ app/Controller/Frontend/ActivityDetail.php
@@ -1,16 +1,16 @@
 <?php
-namespace app\Controllers\Frontend;
+namespace app\Controller\Frontend;
 
 use Silex\Application;
 use Symfony\Component\HttpFoundation\Request;
 
-class JoinToClan
+class ActivityDetail
 {
     public function indexAction(Application $app, Request $request){
 
         return $app['twig']->render('frontend\index.twig', array(
-            'titleAnimation' => "swing",
-            'urlCode' => JOIN_TO_CLAN_URL,
+            'titleAnimation' => "bounce",
+            'urlCode' => ACTIVITY_DETAIL_URL,
         ));
     }
 }
